% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{act_pattern}
\alias{act_pattern}
\title{Single activation pattern example (30 cycles)}
\format{
A data frame of two columns:

\describe{
\item{\code{time}}{Normalised time in points.}\cr
\item{\code{signal}}{Activation pattern}
}
}
\source{
\doi{10.1152/jn.00360.2020}
}
\usage{
act_pattern
}
\description{
A data frame containing one activation pattern extracted from one wild type mouse
walking on a treadmill.
}
\keyword{datasets}
