% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{act_patterns}
\alias{act_patterns}
\title{All activation patterns of one synergy example (30 cycles)}
\format{
A data frame of four columns:

\describe{
\item{\code{time}}{Normalised time in points.}\cr
\item{\code{Syn1}}{Activation pattern of synergy 1}\cr
\item{\code{Syn2}}{Activation pattern 2}\cr
\item{\code{Syn3}}{Activation pattern of synergy 3}
}
}
\source{
\doi{10.1152/jn.00360.2020}
}
\usage{
act_patterns
}
\description{
A demo and incomplete \code{musclesyneRgies} object containing time info and three activation patterns
extracted from one wild type mouse walking on a treadmill.
}
\keyword{datasets}
