% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoA.R
\name{CoA}
\alias{CoA}
\title{Centre of activity}
\usage{
CoA(x, tol_rel = 1e-06)
}
\arguments{
\item{x}{A numeric time series of length n (samples represent bin centres at 1..n).}

\item{tol_rel}{Relative tolerance (0–1). If the mean resultant length
is less than this value, the CoA is considered undefined and \code{NA} is returned.
Default is 1e-6.}
}
\value{
The centre of activity of the time series, calculated with
circular statistics: a numeric scalar in (0, n] giving the centre of activity on the 1..n axis.
}
\description{
Centre of activity
}
\examples{
# Number of users connected to the Internet through a server every minute
ts <- datasets::WWWusage[1:80]

# Calculate CoA
ts_CoA <- CoA(ts)

# Plot
plot(ts, ty = "l", xlab = "Time", ylab = "Number of users")
graphics::abline(v = ts_CoA, lwd = 2, lty = 2)
}
\references{
Martino, G. et al.
Locomotor patterns in cerebellar ataxia.
J. Neurophysiol. 112, 2810–2821 (2014).
}
