% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FILT_EMG}
\alias{FILT_EMG}
\title{Filtered EMG example}
\format{
A list containing one object of class \code{EMG}
with elements \code{cycles} and \code{emg}, both data frames.
\describe{
\item{\code{ID0012_TW_01}}{Object of class \code{EMG} containing the two following data frames:}
\item{\code{cycles}}{Gait cycle-timings, in seconds.}
\item{\code{emg}}{Filtered and time-normalised EMG, first column is time in points, muscles named as:\cr\cr
ME=gluteus medius\cr
MA=gluteus maximus\cr
FL=tensor fasciae latae\cr
RF=rectus femoris\cr
VM=vastus medialis\cr
VL=vastus lateralis\cr
ST=semitendinosus\cr
BF=biceps femoris\cr
TA=tibialis anterior\cr
PL=peroneus longus\cr
GM=gastrocnemius medialis\cr
GL=gastrocnemius lateralis\cr
SO=soleus}
}
}
\source{
\doi{10.1016/j.isci.2019.100796}
}
\usage{
FILT_EMG
}
\description{
A list containing filtered and time-normalised electromyographic (EMG)
human data from the right-side lower limb recorded during one walking trial.
}
\keyword{datasets}
