% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primitive}
\alias{primitive}
\title{Single motor primitive example (30 cycles)}
\format{
A data frame of two columns:

\describe{
\item{\code{time}}{Normalised time in points.}\cr
\item{\code{signal}}{Motor primitive}
}
}
\source{
\url{https://doi.org/10.1152/jn.00360.2020}
}
\usage{
primitive
}
\description{
A data frame containing one motor primitive extracted from one wild type mouse
walking on a treadmill.
}
\keyword{datasets}
