% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primitives}
\alias{primitives}
\title{All motor primitives of one synergy example (30 cycles)}
\format{
A data frame of four columns:

\describe{
\item{\code{time}}{Normalised time in points.}\cr
\item{\code{Syn1}}{Motor primitive of synergy 1}\cr
\item{\code{Syn2}}{Motor primitive of synergy 2}\cr
\item{\code{Syn3}}{Motor primitive of synergy 3}
}
}
\source{
\doi{10.1152/jn.00360.2020}
}
\usage{
primitives
}
\description{
A demo and incomplete \code{musclesyneRgies} object containing time info and three motor primitives
extracted from one wild type mouse walking on a treadmill.
}
\keyword{datasets}
