% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SYNS}
\alias{SYNS}
\title{Muscle synergies example}
\format{
A list containing 15 objects of class \code{musclesyneRgies}, each of which represents
a walking trial from a different person.

\describe{
\item{\code{ID0012_TW_01}\cr
\code{ID0001_TW_01}\cr
\code{ID0002_TW_01}\cr
\code{ID0003_TW_01}\cr
\code{ID0004_TW_01}\cr
\code{ID0005_TW_01}\cr
\code{ID0006_TW_01}\cr
\code{ID0007_TW_01}\cr
\code{ID0008_TW_01}\cr
\code{ID0009_TW_01}\cr
\code{ID0010_TW_01}\cr
\code{ID0011_TW_01}\cr
\code{ID0012_TW_01}\cr
\code{ID0013_TW_01}\cr
\code{ID0014_TW_01}\cr
\code{ID0015_TW_01}\cr}{Objects of class \code{musclesyneRgies} containing the following items:}
\item{\code{syns}}{Factorisation rank or minimum number of synergies.}
\item{\code{M}}{Motor modules (time-invariant coefficients)}
\item{\code{P}}{Motor primitives (time-dependent coefficients)}
\item{\code{V}}{Original data, muscles named as:\cr\cr
ME=gluteus medius\cr
MA=gluteus maximus\cr
FL=tensor fasciae latae\cr
RF=rectus femoris\cr
VM=vastus medialis\cr
VL=vastus lateralis\cr
ST=semitendinosus\cr
BF=biceps femoris\cr
TA=tibialis anterior\cr
PL=peroneus longus\cr
GM=gastrocnemius medialis\cr
GL=gastrocnemius lateralis\cr
SO=soleus}
\item{\code{Vr}}{Reconstructed data, muscles named as in \code{Vr}}
\item{\code{iterations}}{Number of iterations to convergence}
\item{\code{R2}}{Quality of reconstruction (coefficient of determination)}
\item{\code{classification}}{Classification type (e.g., \code{none}, \code{k-means}, \code{NMF}, etc.)}
}
}
\source{
\doi{10.1016/j.isci.2019.100796}
}
\usage{
SYNS
}
\description{
A list created by synsNMF containing muscle synergies extracted from
15 humans walking on a treadmill.
}
\keyword{datasets}
