% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_kmeans.R
\name{classify_kmeans}
\alias{classify_kmeans}
\title{Muscle synergy classification with k-means}
\usage{
classify_kmeans(
  x,
  MSE_lim = 0.001,
  path_for_graphs = NA,
  filetype = "png",
  width = 1620,
  height = 1300,
  resolution = 280,
  interactive = TRUE
)
}
\arguments{
\item{x}{A list of \code{musclesyneRgies} objects}

\item{MSE_lim}{Mean squared error threshold for determining the minimum number of clusters}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}

\item{interactive}{Logical, ask for interactive re-ordering or go full automated?}
}
\value{
List of \code{musclesyneRgies} objects, each with elements:\cr
\itemize{
\item \code{syns} factorisation rank or minimum number of synergies\cr
\item \code{M} motor modules (time-invariant coefficients)\cr
\item \code{P} motor primitives (time-dependent coefficients)\cr
\item \code{V} original data\cr
\item \code{Vr} reconstructed data\cr
\item \code{iterations} number of iterations to convergence\cr
\item \code{R2} quality of reconstruction (coefficient of determination)
\item \code{rank_type} was the rank \code{fixed} or \code{variable}?\cr
\item \code{classification} classification type (\code{k-means})
}
}
\description{
Muscle synergy classification with k-means
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path <- getwd()
data_path <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings = FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data("SYNS")

SYNS_classified <- classify_kmeans(SYNS,
  path_for_graphs = path_for_graphs,
  interactive = FALSE
)

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive = TRUE)
}
