% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classified_syns.R
\name{plot_classified_syns}
\alias{plot_classified_syns}
\title{Plot muscle synergies}
\usage{
plot_classified_syns(
  x,
  path_for_graphs = NA,
  condition,
  filetype = "png",
  width = 1800,
  height = 625,
  resolution = 280
)
}
\arguments{
\item{x}{List of objects of class \code{musclesyneRgies} (must be classified)}

\item{path_for_graphs}{Path where plots should be saved}

\item{condition}{Character: the condition that is being analysed, for archiving purposes}

\item{filetype}{Plot file type}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
Global plot containing the average classified muscle synergies and individual trials
(motor modules) or standard deviations (motor primitives)
}
\description{
Plot muscle synergies
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path <- getwd()
data_path <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings = FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(SYNS)

## Classify synergies with k-means
SYNS_classified <- classify_kmeans(SYNS,
  path_for_graphs = path_for_graphs,
  interactive = FALSE
)

## Export plots
plot_classified_syns(SYNS_classified,
  path_for_graphs = path_for_graphs,
  condition = "TW"
)

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive = TRUE)
}
