% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classified_syns_UMAP.R
\name{plot_classified_syns_UMAP}
\alias{plot_classified_syns_UMAP}
\title{Plot 2D UMAP of muscle synergies}
\usage{
plot_classified_syns_UMAP(
  x,
  path_for_graphs,
  condition,
  filetype = "png",
  width = 1800,
  height = 2840,
  resolution = 280
)
}
\arguments{
\item{x}{List of objects of class \code{musclesyneRgies} (must be classified)}

\item{path_for_graphs}{Path where plots should be saved}

\item{condition}{Character: the condition that is being analysed, for archiving purposes}

\item{filetype}{Plot file type}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
2D UMAP plot of classified synergies.
}
\description{
Plot 2D UMAP of muscle synergies
}
\examples{
## Create "Graphs" folder if it does not exist
data_path  <- getwd()
data_path  <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings=FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(SYNS)

## Classify synergies with k-means
SYNS_classified <- classify_kmeans(SYNS,
                                   path_for_graphs=path_for_graphs,
                                   interactive=FALSE)

## Export plots
plot_classified_syns_UMAP(SYNS_classified,
                          path_for_graphs=path_for_graphs,
                          condition="TW")

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive=TRUE)
}
