% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meanEMG.R
\name{plot_meanEMG}
\alias{plot_meanEMG}
\title{Plot EMG averaged across all cycles}
\usage{
plot_meanEMG(
  x,
  trial,
  row_number = NA,
  col_number = 1,
  dark_mode = FALSE,
  line_size = 0.6,
  line_col = "black",
  path_for_graphs = NA,
  filetype,
  width,
  height,
  resolution
)
}
\arguments{
\item{x}{A data frame containing filtered EMG organised in columns}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{row_number}{How many rows should the final plot be divided into?}

\item{col_number}{How many columns should the final plot be divided into?}

\item{dark_mode}{To enable dark mode}

\item{line_size}{Line thickness}

\item{line_col}{Line colour}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type (e.g., "png" or "svg")}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
Exports average filtered and normalised EMG.
}
\description{
Plot EMG averaged across all cycles
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path <- getwd()
data_path <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings = FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(FILT_EMG)

for (ii in seq_along(FILT_EMG)) {
  ## Plot the average filtered and time-normalised EMG
  plot_meanEMG(FILT_EMG[[ii]],
    trial = names(FILT_EMG)[ii],
    path_for_graphs = path_for_graphs,
    row_number = 4,
    col_number = 4,
    dark_mode = TRUE,
    line_col = "tomato3",
    filetype = "png",
    width = 2000,
    height = 2000,
    resolution = 230
  )
}

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive = TRUE)
}
