% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meanEMG.R
\name{plot_meanEMG}
\alias{plot_meanEMG}
\title{Plot EMG averaged across all cycles}
\usage{
plot_meanEMG(
  x,
  trial,
  row_number = NA,
  col_number = 1,
  dark_mode = FALSE,
  line_size = 0.6,
  line_col = "black",
  path_for_graphs = NA,
  filetype,
  width,
  height,
  resolution
)
}
\arguments{
\item{x}{A data frame containing filtered EMG organised in columns}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{row_number}{How many rows should the final plot be divided into?}

\item{col_number}{How many columns should the final plot be divided into?}

\item{dark_mode}{To enable dark mode}

\item{line_size}{Line thickness}

\item{line_col}{Line colour}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type (e.g., "png" or "svg")}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
Exports average filtered and normalised EMG.
}
\description{
Plot EMG averaged across all cycles
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
# Load some data
data(FILT_EMG)

# Plot
for (ii in seq_along(FILT_EMG)) {
  # Plot the average filtered and time-normalised EMG
  plot_meanEMG(FILT_EMG[[ii]],
    trial = names(FILT_EMG)[ii],
    row_number = 4,
    col_number = 4,
    dark_mode = TRUE,
    line_col = "tomato3"
  )
}
}
