% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meanEMG.R
\name{plot_meanEMG}
\alias{plot_meanEMG}
\title{Plot EMG averaged across all cycles}
\usage{
plot_meanEMG(
  x,
  trial,
  path_for_graphs = NA,
  filetype = "png",
  width = 2000,
  height = 1875,
  resolution = 140
)
}
\arguments{
\item{x}{A data frame containing filtered EMG organised in columns}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
Exports average filtered and normalised EMG.
}
\description{
Plot EMG averaged across all cycles
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path <- getwd()
data_path <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings = FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(FILT_EMG)

for (ii in seq_along(FILT_EMG)) {
  ## Plot the average filtered and time-normalised EMG
  plot_meanEMG(FILT_EMG[[ii]],
    trial = names(FILT_EMG)[ii],
    path_for_graphs = path_for_graphs
  )
}

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive = TRUE)
}
