% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rawEMG.R
\name{plot_rawEMG}
\alias{plot_rawEMG}
\title{Plot raw EMG}
\usage{
plot_rawEMG(
  x,
  trial,
  plot_time = 3,
  start = 1,
  row_number = NA,
  col_number = 1,
  dark_mode = FALSE,
  line_size = 0.3,
  line_col = "black",
  path_for_graphs = NA,
  filetype,
  width,
  height,
  resolution
)
}
\arguments{
\item{x}{Object of class \code{EMG} with elements \code{cycles} and \code{emg}}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{plot_time}{How many seconds of data should be plotted?}

\item{start}{At which data point should the plot start?}

\item{row_number}{How many rows should the final plot be divided into?}

\item{col_number}{How many columns should the final plot be divided into?}

\item{dark_mode}{To enable dark mode}

\item{line_size}{Line thickness}

\item{line_col}{Line colour}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type (e.g., "png" or "svg")}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels, for each muscle}

\item{resolution}{Plot resolution in pixels}
}
\value{
Exports raw EMG plots of the specified length.
}
\description{
Plot raw EMG
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path <- getwd()
data_path <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings = FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(RAW_DATA)

for (ii in seq_along(RAW_DATA)) {
  ## Plot the first three seconds of each trial in dark mode
  plot_rawEMG(RAW_DATA[[ii]],
    trial = names(RAW_DATA)[ii],
    path_for_graphs = path_for_graphs,
    row_number = 4,
    col_number = 4,
    dark_mode = TRUE,
    line_col = "tomato3",
    filetype = "png",
    width = 2000,
    height = 2000,
    resolution = 230
  )
}

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive = TRUE)
}
