% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rawEMG.R
\name{plot_rawEMG}
\alias{plot_rawEMG}
\title{Plot raw EMG}
\usage{
plot_rawEMG(
  x,
  trial,
  plot_time = 3,
  start = 1,
  path_for_graphs,
  filetype = "png",
  width = 2000,
  height = 500,
  resolution = 280
)
}
\arguments{
\item{x}{Object of class \code{EMG} with elements \code{cycles} and \code{emg}}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{plot_time}{How many seconds of data should be plotted?}

\item{start}{At which data point should the plot start?}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels, for each muscle}

\item{resolution}{Plot resolution in pixels}
}
\value{
Exports raw EMG plots of the specified length.
}
\description{
Plot raw EMG
}
\examples{
## Create "Graphs" folder if it does not exist
data_path  <- getwd()
data_path  <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings=FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(RAW_DATA)

for (ii in seq_along(RAW_DATA)) {
  ## Plot the first three seconds of each trial
  plot_rawEMG(RAW_DATA[[ii]],
              trial=names(RAW_DATA)[ii],
              path_for_graphs=path_for_graphs)
}

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive=TRUE)
}
