% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_syn_trials.R
\name{plot_syn_trials}
\alias{plot_syn_trials}
\title{Plot muscle synergies (individual trials)}
\usage{
plot_syn_trials(
  x,
  max_syns,
  trial,
  dark_mode = FALSE,
  line_size = 0.6,
  line_col = "black",
  sd_col = "grey80",
  path_for_graphs = NA,
  filetype,
  width,
  height,
  resolution
)
}
\arguments{
\item{x}{Object of class \code{musclesyneRgies}}

\item{max_syns}{Number of synergies to be plotted or how many rows should the final panel
be divided into}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{dark_mode}{To enable dark mode}

\item{line_size}{Line thickness}

\item{line_col}{Line colour}

\item{sd_col}{Standard deviation ribbon colour}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type (e.g., "png" or "svg")}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
Plots of the unclassified synergies, trial by trial.
}
\description{
Plot muscle synergies (individual trials)
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path <- getwd()
data_path <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings = FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(FILT_EMG)
SYNS <- pbapply::pblapply(FILT_EMG, synsNMF)

max_syns <- max(unlist(lapply(SYNS, function(x) x$syns)))

## Plot the first (and only) trial in the list
plot_syn_trials(SYNS[[1]],
  max_syns = max_syns,
  trial = names(SYNS)[1],
  path_for_graphs = path_for_graphs,
  dark_mode = TRUE,
  line_size = 0.8,
  line_col = "tomato1",
  sd_col = "tomato4",
  filetype = "png",
  width = 2000,
  height = 500 * max_syns,
  resolution = 280
)

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive = TRUE)
}
