% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_syn_trials.R
\name{plot_syn_trials}
\alias{plot_syn_trials}
\title{Plot muscle synergies (individual trials)}
\usage{
plot_syn_trials(
  x,
  max_syns,
  trial,
  path_for_graphs = NA,
  filetype = "png",
  width = 2000,
  height = 500 * max_syns,
  resolution = 280
)
}
\arguments{
\item{x}{Object of class \code{musclesyneRgies}}

\item{max_syns}{Maximum number of synergies to be plotted}

\item{trial}{Character: the name of the considered trial, for archiving purposes}

\item{path_for_graphs}{Path where plots should be saved}

\item{filetype}{Plot file type}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{resolution}{Plot resolution in pixels}
}
\value{
Plots of the unclassified synergies, trial by trial.
}
\description{
Plot muscle synergies (individual trials)
}
\details{
If \code{path_for_graphs} is not specified, plots will appear in the plot pane and will not be saved.
}
\examples{
## Create "Graphs" folder if it does not exist
data_path  <- getwd()
data_path  <- paste0(data_path, .Platform$file.sep)
dir.create("Graphs", showWarnings=FALSE)
path_for_graphs <- paste0(data_path, "Graphs", .Platform$file.sep)

data(SYNS)

max_syns <- max(unlist(lapply(SYNS, function(x) x$syns)))

## Plot the first two trials in the list
for (ii in c(1, 2)) {
  ## Plot the unclassified synergies, trial-by-trial
  plot_syn_trials(SYNS[[ii]],
                  max_syns=max_syns,
                  trial=names(SYNS)[ii],
                  path_for_graphs=path_for_graphs)
}

## Check plots in the new folder before running the following (will delete!)

## Delete folder
unlink("Graphs", recursive=TRUE)
}
